+++ front/lib/assistant.ts
@@ -1,9 +1,11 @@
 import type {
-  AgentUsageType,
+  DataSourceType,
   LightAgentConfigurationType,
 } from "@dust-tt/types";
 import type { SupportedModel } from "@dust-tt/types";
 import { SUPPORTED_MODEL_CONFIGS } from "@dust-tt/types";
+
+import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
 
 export function isLargeModel(model: unknown): model is SupportedModel {
   const maybeSupportedModel = model as SupportedModel;
@@ -99,24 +101,40 @@ export function compareAgentsForSort(
   return 0; // Default: keep the original order
 }
 
-export function assistantUsageMessage({
-  usage,
-  isLoading,
-  isError,
-}: {
-  usage: AgentUsageType | null;
-  isLoading: boolean;
-  isError: boolean;
-}) {
-  if (isError) {
-    return "Error loading usage data.";
-  } else if (isLoading) {
-    return "Loading usage data...";
-  } else if (usage) {
-    return `This assistant has been used by ${usage.userCount} ${
-      usage.userCount > 1 ? "people" : "person"
-    } in ${usage.messageCount} message${
-      usage.messageCount > 1 ? "s" : ""
-    } over the last ${usage.timePeriodSec / (60 * 60 * 24)} days.`;
-  }
+// Order in the following format : connectorProvider > empty > webcrawler
+export function orderDatasourceByImportance(dataSources: DataSourceType[]) {
+  return dataSources.sort((a, b) => {
+    const aConnector = a.connectorProvider;
+    const bConnector = b.connectorProvider;
+
+    const order = Object.keys(CONNECTOR_CONFIGURATIONS)
+      .filter(
+        (key) =>
+          CONNECTOR_CONFIGURATIONS[key as keyof typeof CONNECTOR_CONFIGURATIONS]
+            .connectorProvider !==
+          CONNECTOR_CONFIGURATIONS.webcrawler.connectorProvider
+      )
+      .map(
+        (key) =>
+          CONNECTOR_CONFIGURATIONS[key as keyof typeof CONNECTOR_CONFIGURATIONS]
+            .connectorProvider
+      );
+
+    if (aConnector === CONNECTOR_CONFIGURATIONS.webcrawler.connectorProvider) {
+      return 1;
+    }
+
+    if (bConnector === CONNECTOR_CONFIGURATIONS.webcrawler.connectorProvider) {
+      return -1;
+    }
+
+    const indexA = aConnector ? order.indexOf(aConnector) : order.length;
+    const indexB = bConnector ? order.indexOf(bConnector) : order.length;
+
+    if (indexA === -1 && indexB === -1) {
+      return 0;
+    }
+
+    return indexA - indexB;
+  });
 }
